Add process raw data buffer function and renamed a lot of mission critical code
