+++ changelog.md
@@ -1,3 +1,15 @@
+# 0.0.2
+
+### Breaking Changes
+
+* Renamed `Sample` to `Utilities` in `index.js`
+* Renamed openBCIUtilities.js to openBCIUtilities.js
+* Renamed openBCIUtilities-test.js to openBCIUtilities-test.js
+
+### New Features
+
+* Added a function in the sample module that parses a raw buffer of OpenBCI data, extracts raw data packets and returns the buffer with just the raw data packets removed. Allowing the user to process other data that is not a raw data.
+
 # 0.0.1
 
 Initial release
