+++ index.js
@@ -1,3 +1,3 @@
 module.exports.Constants = require('./openBCIConstants');
 module.exports.Debug = require('./openBCIDebug');
-module.exports.Sample = require('./openBCISample');
+module.exports.Utilities = require('./openBCIUtilities');
